debounce buggy, revert back to normal submit, save favorite to localStorage
